updated to work with cucumber v8 and webdriverio v8
