+++ runtime/accessibility/accessibilityLib.js
@@ -18,9 +18,11 @@
  SOFTWARE.
  */
 const axe = require('axe-core');
+const fs = require('fs-extra');
+const path = require('path');
 
 let errorCount = 0;
-let totalErrorCout = 0;
+let totalErrorCount = 0;
 
 const envName = global.env.envName.toLowerCase();
 
@@ -29,7 +31,7 @@ module.exports = {
     return errorCount;
   },
   getAccessibilityTotalError() {
-    return totalErrorCout;
+    return totalErrorCount;
   },
   /** main method to be called for accessibility report */
   async getAccessibilityReport(PageName) {
@@ -62,14 +64,10 @@ module.exports = {
     }
 
     errorCount = results.violations.length + results.incomplete.length;
-    totalErrorCout += errorCount;
+    totalErrorCount += errorCount;
   },
 
   async GenerateFinalAccessibilityReport(fullData, additionalData, Pagename, browserName) {
-    // eslint-disable-next-line global-require
-    const fs = require('fs-extra');
-    // eslint-disable-next-line global-require
-    const path = require('path');
     const sample = fs.readFileSync(path.resolve(__dirname, './ReportSample'), 'utf-8');
 
     const addDataInHtml = sample.replace('XXX-DetailData', JSON.stringify(fullData));
@@ -104,10 +102,6 @@ module.exports = {
   },
 
   async GenerateFinalAccessibilityLiteReport(fullData, additionalData, Pagename, browserName) {
-    // eslint-disable-next-line global-require
-    const fs = require('fs-extra');
-    // eslint-disable-next-line global-require
-    const path = require('path');
     const sample = fs.readFileSync(path.resolve(__dirname, './ReportSample-lite'), 'utf-8');
     const summaryViewInfo = await module.exports.htmlsummaryView(fullData);
     const summaryViewData = await module.exports.htmlsummaryData(fullData.testEnvironment, fullData);
@@ -148,93 +142,6 @@ module.exports = {
     );
   },
 
-  /** This is depricated */
-  GenerateAccessibilityReport(detailreport, Pagename) {
-    // eslint-disable-next-line global-require
-    const fs = require('fs-extra');
-    const sample = fs.readFileSync('./Accessibility/ReportSample-lite', 'utf-8');
-    const violationHtml = module.exports.CreateHTMLBase(detailreport.violations, 1000);
-    const incompleteHtml = module.exports.CreateHTMLBase(detailreport.incomplete, 2000);
-    const passedHtml = module.exports.CreateHTMLBase(detailreport.passes, 3000);
-    const iapplicableHtml = module.exports.CreateHTMLBase(detailreport.inapplicable, 4000);
-    let NewHtml = sample.replace('XXX-Violation', violationHtml);
-    NewHtml = NewHtml.replace('XXX-Incomplete', incompleteHtml);
-    NewHtml = NewHtml.replace('XXX-Passed', passedHtml);
-    NewHtml = NewHtml.replace('XXX-Inapplicable', iapplicableHtml);
-    const dirAcc = `${global.paths.reports}/${envName}/Accessibility`;
-    if (!fs.existsSync(dirAcc)) {
-      fs.mkdirSync(dirAcc);
-    }
-
-    fs.writeFile(`${dirAcc}/AccessbilityReport_${Pagename}.html`, NewHtml, 'utf-8', (err) => {
-      if (err) throw err;
-    });
-  },
-
-  CreateHTMLBase(data, y) {
-    let txt = '';
-    // var y=0;
-    let z = 0;
-    // eslint-disable-next-line guard-for-in,no-restricted-syntax
-    for (x in data) {
-      // eslint-disable-next-line no-param-reassign
-      y += 1;
-      z += 1;
-      txt += '<div class="panel panel-default">';
-      txt += '<div class="panel-heading" style="background-color:#F9E79F">';
-      txt += '<h4 class="panel-title">';
-
-      txt += '<a data-toggle="collapse" data-parent="#accordion" href="#collapse';
-      txt += y;
-      txt += '">';
-      txt += `${z}: ${data[x].help.replace(/(<([^>]+)>)/gi, '')}`;
-      // txt+= z+": "+(data[x].help).replace( /<>/ig, '');
-      txt += '</a>';
-
-      txt += '<div style="float: right">';
-      txt += data[x].impact;
-      txt += '</div>';
-      txt += '</h4>';
-      txt += '</div>';
-      txt += '<div id="collapse';
-      txt += y;
-      txt += '" class="panel-collapse collapse ">';
-      txt += '<div class="panel-body" style="background-color: #92a8d1">';
-      txt += '<a href=';
-      txt += data[x].helpUrl;
-      txt += ' target="_blank">';
-
-      txt += `<xmp>${data[x].help}</xmp>`;
-      txt += '</a>';
-      txt += '<div>';
-      txt += `<xmp>${data[x].description}</xmp>`;
-      txt += '</div>';
-      txt += '<div>';
-      txt += data[x].impact;
-      txt += '</div>';
-      txt += '<div style="overflow:auto">';
-      txt += '<xmp>';
-      txt += JSON.stringify(data[x].nodes, undefined, 2);
-      txt += '</xmp>';
-      txt += '</div>';
-      txt += '</div>';
-
-      txt += '</div>';
-
-      txt += '</div>';
-    }
-    return txt;
-  },
-
-  /** fuction to attach screen shot into the acceesibility report */
-  async takeScreenPic() {
-    let picstring;
-    await browser.takeScreenshot().then((screenShot) => {
-      picstring = screenShot;
-    });
-    return picstring;
-  },
-
   /** to get current data time */
   getCurrentDateTime() {
     const today = new Date();
@@ -266,7 +173,6 @@ module.exports = {
       changeColor = true;
     }
     let txt = '';
-    // var y=0;
     let z = 0;
     // eslint-disable-next-line guard-for-in,no-restricted-syntax
     for (const x in data) {
