+++ runtime/chromeDriver.js
@@ -94,7 +94,7 @@ module.exports = async function chromeDriver(options) {
   global.browser = await remote(extendedOptions);
   if (isUTAMTest) {
     const utamInstance = new UtamWdioService(utamConfig, extendedOptions.capabilities, extendedOptions);
-    await utamInstance.before(extendedOptions.capabilities);
+    await utamInstance.before(extendedOptions.capabilities, null, browser);
   }
   await browser.setWindowSize(1280, 1024);
   return browser;
