+++ step_definitions/google-search-steps.js
@@ -9,11 +9,13 @@ try {
 Given(/^The user navigates to Google$/, async () => {
   await helpers.loadPage(env.web_utam_url, 10);
   const modal = await browser.$('button[id="L2AGLb"]');
-  if (await modal.isDisplayed()) {
+  if ((await modal.isDisplayed()) === true) {
     const google = await utam.load(Google);
     const acceptModalBtn = await google.getAcceptModalBtn();
     await acceptModalBtn.click();
-    await google.waitForAbsence();
+    if (process.platform !== 'darwin') {
+      await google.waitForAbsence();
+    }
   }
 });
 When(/^They use the engine to search for a word$/, async () => {
