+++ step_definitions/search-steps.js
@@ -14,11 +14,6 @@ Then(/^they should see some results$/, async () => {
   await pageObjects.search.searchResult();
 });
 
-When(/^The screenshots should differ with the filenames "([^"]*)" & "([^"]*)"$/, async (fileName1, fileName2) => {
-  const numDiffPixels = await helpers.imagePixelMatch(fileName1, fileName2);
-  assert.isAbove(numDiffPixels, 0, 'Num of pixels should be greater than 0');
-});
-
 Then(/^Add modHeader (.*) (.*) (.*)$/, async (extName, username, password) => {
   if (browserName === 'chrome') {
     await helpers.modHeader(extName, username, password);
